introaboutme/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
future==0.18.2
Pillow==8.3.1
pytz==2021.1
serializers==0.2.4
sqlparse==0.4.1
